refactor(cautils): lift workload identifier parsing out of cmd/scan
