+++ cmd/scan/workload.go
@@ -1,11 +1,8 @@
 package scan
 
 import (
-	"errors"
 	"fmt"
 	"io"
-	"regexp"
-	"strings"
 
 	"github.com/kubescape/kubescape/v4/cmd/shared"
 	"github.com/kubescape/kubescape/v4/core/cautils"
@@ -41,8 +38,20 @@ var (
 
 
 `, cautils.ExecName())
+)
 
-	ErrInvalidWorkloadIdentifier = errors.New("invalid workload identifier, expected <kind>[.<version>[.<group>]]/<name>")
+// The workload identifier grammar is shared with the MCP server, which cannot
+// import cmd/scan without pulling in cobra and the whole command tree, so it
+// lives in core/cautils/workloadidentifier.go. These aliases keep the call
+// sites below unchanged.
+//
+// ErrInvalidWorkloadIdentifier is aliased rather than redeclared on purpose:
+// callers compare against it by identity, which a second errors.New carrying
+// the same message would not satisfy.
+var (
+	ErrInvalidWorkloadIdentifier  = cautils.ErrInvalidWorkloadIdentifier
+	parseWorkloadIdentifierString = cautils.ParseWorkloadIdentifierString
+	validateWorkloadIdentifier    = cautils.ValidateWorkloadIdentifier
 )
 
 // controlCmd represents the control command
@@ -185,63 +194,4 @@ func setWorkloadScanInfo(scanInfo *cautils.ScanInfo, kind string, name string, a
 	}
 
 	return policyIdentifiers
-}
-
-func validateWorkloadIdentifier(workloadIdentifier string) error {
-	_, _, _, _, err := parseWorkloadIdentifierString(workloadIdentifier)
-	return err
-}
-
-func parseWorkloadIdentifierString(workloadIdentifier string) (namespace, kind, name, apiVersion string, err error) {
-	// workloadIdentifier is in the form of kind/name or namespace/kind/name
-	// example: default/Deployment/nginx-deployment
-	x := strings.Split(workloadIdentifier, "/")
-	if len(x) == 2 {
-		if x[0] == "" || x[1] == "" {
-			return "", "", "", "", ErrInvalidWorkloadIdentifier
-		}
-		parsedKind, parsedApiVersion, err := parseKindAndApiVersion(x[0])
-		if err != nil {
-			return "", "", "", "", err
-		}
-		return "", parsedKind, x[1], parsedApiVersion, nil
-	}
-	if len(x) == 3 {
-		if x[0] == "" || x[1] == "" || x[2] == "" {
-			return "", "", "", "", ErrInvalidWorkloadIdentifier
-		}
-		parsedKind, parsedApiVersion, err := parseKindAndApiVersion(x[1])
-		if err != nil {
-			return "", "", "", "", err
-		}
-		return x[0], parsedKind, x[2], parsedApiVersion, nil
-	}
-
-	return "", "", "", "", ErrInvalidWorkloadIdentifier
-}
-
-var apiVersionPattern = regexp.MustCompile(`^v\d+((alpha|beta)\d+)?$`)
-
-func parseKindAndApiVersion(kindStr string) (kind, apiVersion string, err error) {
-	parts := strings.Split(kindStr, ".")
-	if len(parts) == 1 {
-		return kindStr, "", nil
-	}
-
-	// Reject empty components
-	for _, part := range parts {
-		if part == "" {
-			return "", "", fmt.Errorf("%w: empty component in %q", ErrInvalidWorkloadIdentifier, kindStr)
-		}
-	}
-
-	if !apiVersionPattern.MatchString(parts[1]) {
-		return "", "", fmt.Errorf("%w: %q is not a valid API version in %q", ErrInvalidWorkloadIdentifier, parts[1], kindStr)
-	}
-
-	if len(parts) >= 3 {
-		group := strings.Join(parts[2:], ".")
-		return parts[0], group + "/" + parts[1], nil // kind.version.group -> group/version
-	}
-	return parts[0], parts[1], nil // kind.version -> version
 }
