+++ core/cautils/workloadidentifier_test.go
@@ -1,0 +1,152 @@
+package cautils
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestParseWorkloadIdentifierString_Invalid(t *testing.T) {
+	tests := []struct {
+		name  string
+		input string
+	}{
+		{
+			name:  "empty identifier",
+			input: "",
+		},
+		{
+			name:  "too many segments",
+			input: "cluster/default/Deployment/nginx",
+		},
+		{
+			name:  "empty segment",
+			input: "default//nginx",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			_, _, _, _, err := ParseWorkloadIdentifierString(tt.input)
+			assert.Error(t, err)
+		})
+	}
+}
+
+func TestParseWorkloadIdentifierString_Valid(t *testing.T) {
+	t.Run("valid identifier", func(t *testing.T) {
+		namespace, kind, name, apiVersion, err := ParseWorkloadIdentifierString("default/Deployment/nginx-deployment")
+		assert.NoError(t, err)
+		assert.Equal(t, "default", namespace)
+		assert.Equal(t, "Deployment", kind)
+		assert.Equal(t, "nginx-deployment", name)
+		assert.Equal(t, "", apiVersion)
+	})
+}
+
+func TestParseWorkloadIdentifierString_Values(t *testing.T) {
+	testCases := []struct {
+		Description    string
+		Input          string
+		WantNamespace  string
+		WantKind       string
+		WantName       string
+		WantApiVersion string
+		WantErr        bool
+	}{
+		{
+			Description:    "valid kind and name",
+			Input:          "Deployment/nginx",
+			WantNamespace:  "",
+			WantKind:       "Deployment",
+			WantName:       "nginx",
+			WantApiVersion: "",
+			WantErr:        false,
+		},
+		{
+			Description:    "valid namespace kind and name",
+			Input:          "default/Deployment/nginx",
+			WantNamespace:  "default",
+			WantKind:       "Deployment",
+			WantName:       "nginx",
+			WantApiVersion: "",
+			WantErr:        false,
+		},
+		{
+			Description:    "valid kind.version and name",
+			Input:          "Pod.v1/nginx",
+			WantNamespace:  "",
+			WantKind:       "Pod",
+			WantName:       "nginx",
+			WantApiVersion: "v1",
+			WantErr:        false,
+		},
+		{
+			Description:    "valid kind.version.group and name",
+			Input:          "Deployment.v1.apps/nginx",
+			WantNamespace:  "",
+			WantKind:       "Deployment",
+			WantName:       "nginx",
+			WantApiVersion: "apps/v1",
+			WantErr:        false,
+		},
+		{
+			Description:    "valid namespace kind.version.group and name",
+			Input:          "default/Deployment.v1.apps/nginx",
+			WantNamespace:  "default",
+			WantKind:       "Deployment",
+			WantName:       "nginx",
+			WantApiVersion: "apps/v1",
+			WantErr:        false,
+		},
+		{
+			Description:    "valid multi-label group",
+			Input:          "Ingress.v1.networking.k8s.io/name",
+			WantNamespace:  "",
+			WantKind:       "Ingress",
+			WantName:       "name",
+			WantApiVersion: "networking.k8s.io/v1",
+			WantErr:        false,
+		},
+		{
+			Description: "invalid empty dotted component",
+			Input:       "Deployment..apps/nginx",
+			WantErr:     true,
+		},
+		{
+			Description: "invalid empty trailing component",
+			Input:       "Deployment./nginx",
+			WantErr:     true,
+		},
+		{
+			Description: "invalid missing apiVersion",
+			Input:       "Deployment.apps/nginx",
+			WantErr:     true,
+		},
+		{
+			Description: "invalid apiVersion segment",
+			Input:       "Deployment.bogus/nginx",
+			WantErr:     true,
+		},
+		{
+			Description: "too many segments",
+			Input:       "cluster/default/Deployment/nginx",
+			WantErr:     true,
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.Description, func(t *testing.T) {
+			namespace, kind, name, apiVersion, err := ParseWorkloadIdentifierString(tc.Input)
+			if tc.WantErr {
+				assert.Error(t, err)
+				return
+			}
+			assert.NoError(t, err)
+			assert.Equal(t, tc.WantNamespace, namespace)
+			assert.Equal(t, tc.WantKind, kind)
+			assert.Equal(t, tc.WantName, name)
+			assert.Equal(t, tc.WantApiVersion, apiVersion)
+		})
+	}
+}
